+++ src/components/modals/human-card/index.js
@@ -23,3 +23,5 @@ export { useTrustedOwnerLinks } from "./useTrustedOwnerLinks.js";
 export { usePendingSignupLink } from "./usePendingSignupLink.js";
 export { useHumanRemoval } from "./useHumanRemoval.js";
 export { SignupLinkDialogs } from "./SignupLinkDialogs.jsx";
+export { HumanCardConfirmDialogs } from "./HumanCardConfirmDialogs.jsx";
+export { MobileNewBookingBar } from "./MobileNewBookingBar.jsx";